+++ experimental/traffic-portal/src/app/api/index.ts
@@ -23,7 +23,6 @@ import { CDNService } from "./cdn.service";
 import { DeliveryServiceService } from "./delivery-service.service";
 import { InvalidationJobService } from "./invalidation-job.service";
 import { MiscAPIsService } from "./misc-apis.service";
-import { ParameterService } from "./parameter.service";
 import { PhysicalLocationService } from "./physical-location.service";
 import { ProfileService } from "./profile.service";
 import { ServerService } from "./server.service";
@@ -36,7 +35,6 @@ export * from "./change-logs.service";
 export * from "./delivery-service.service";
 export * from "./invalidation-job.service";
 export * from "./misc-apis.service";
-export * from "./parameter.service";
 export * from "./physical-location.service";
 export * from "./profile.service";
 export * from "./server.service";
@@ -58,7 +56,6 @@ export * from "./user.service";
 		DeliveryServiceService,
 		InvalidationJobService,
 		MiscAPIsService,
-		ParameterService,
 		PhysicalLocationService,
 		ProfileService,
 		ServerService,
